+++ src/View/Antlers/Parser.php
@@ -569,8 +569,10 @@ class Parser
             // a callback. If it's a query builder instance, we want to use the Query tag's index
             // method to handle the logic. We'll pass the builder into the builder parameter.
             if (isset($data[$name])) {
-                if ($data[$name] instanceof Builder) {
-                    $parameters['builder'] = $data[$name];
+                $value = $data[$name];
+                $value = $value instanceof Value ? $value->value() : $value;
+                if ($value instanceof Builder) {
+                    $parameters['builder'] = $value;
                     $name = 'query';
                 }
             }
@@ -1233,6 +1235,10 @@ class Parser
 
         if ($context instanceof Value) {
             $context = $context->value();
+        }
+
+        if ($context instanceof Builder) {
+            $context = $context->get();
         }
 
         if ($context instanceof Augmentable) {